Separate to Ui, Storage, Parser, TaskList, Add files into relevant packages
