+++ data/command.txt
@@ -1,1 +1,3 @@
-D | 1 | return book | 2/12/2019 1800
+T | 0 | a
+T | 0 | b
+T | 0 | a
