+++ src/main/java/duke/ChatBot.java
@@ -1,11 +1,11 @@
-package main.java;
+package main.java.duke;
 
-import main.java.data.Storage;
-import main.java.data.TaskList;
-import main.java.dukeexceptions.DukeException;
-import main.java.parser.Parser;
-import main.java.responses.Response;
-import main.java.ui.Ui;
+import main.java.duke.data.Storage;
+import main.java.duke.data.TaskList;
+import main.java.duke.dukeexceptions.ForeignException;
+import main.java.duke.parser.Parser;
+import main.java.duke.responses.Response;
+import main.java.duke.ui.Ui;
 
 /**
  * This is the chatbot object.
@@ -34,16 +34,18 @@ public class ChatBot {
       while(isRunning) {
         try {
           String nextLine = cmdLine.getNextLine();
-          Command nextCommand = cmdLine.getCommand(nextLine);
+          Command nextCommand = null;
+
+          nextCommand = cmdLine.getCommand(nextLine);
           Response nextResponse = commandHandler.getResponse(nextCommand, nextLine,store, taskList);
           Ui.callResponse(nextResponse);
           if (nextCommand == Command.bye) {
             break;
           }
-        } catch (DukeException e) {
-          e.callback();
+        } catch (ForeignException e) {
+        e.printStackTrace();
         }
-    }
+      }
 
   }
