+++ src/main/java/duke/Command.java
@@ -1,4 +1,4 @@
-package main.java;
+package main.java.duke;
 
 public enum Command {
