+++ src/main/java/duke/Duke.java
@@ -1,4 +1,4 @@
-package main.java;
+package main.java.duke;
 
 public class Duke {
   public static void main(String[] args) {
