+++ src/main/java/duke/data/TaskList.java
@@ -1,6 +1,6 @@
 package main.java.duke.data;
 
-import main.java.task.Task;
+import main.java.duke.task.Task;
 
 import java.util.ArrayList;
