+++ src/main/java/duke/dukeexceptions/DukeException.java
@@ -1,4 +1,4 @@
-package main.java.dukeexceptions;
+package main.java.duke.dukeexceptions;
 
 public class DukeException extends  Exception{
     DukeException(String msg) {
