+++ src/main/java/duke/dukeexceptions/ForeignException.java
@@ -1,7 +1,7 @@
-package main.java.dukeexceptions;
+package main.java.duke.dukeexceptions;
 
 
-public class ForeignException extends DukeException {
+public class ForeignException extends main.java.duke.dukeexceptions.DukeException {
     public ForeignException(String msg) {
         super(msg);
     }
