+++ src/main/java/duke/responses/MarkResponse.java
@@ -1,6 +1,6 @@
-package main.java.responses;
+package main.java.duke.responses;
 
-import main.java.task.Task;
+import main.java.duke.task.Task;
 
 public class MarkResponse implements Response {
     Task currTask;
