+++ src/main/java/duke/responses/UnMarkResponse.java
@@ -1,7 +1,7 @@
-package main.java.responses;
+package main.java.duke.responses;
 
 
-import main.java.task.Task;
+import main.java.duke.task.Task;
 
 public class UnMarkResponse implements Response {
     Task currTask;
