+++ src/main/java/duke/task/Event.java
@@ -1,4 +1,4 @@
-package main.java.task;
+package main.java.duke.task;
 
 import java.time.LocalDateTime;
