+++ src/main/java/duke/task/Task.java
@@ -1,4 +1,4 @@
-package main.java.task;
+package main.java.duke.task;
 
 import java.time.format.DateTimeFormatter;
